+++ include/ippp/core/sampling/GaussianSampling.hpp
@@ -33,7 +33,7 @@ class GaussianSampling : public Sampling<dim> {
   public:
     GaussianSampling(const std::shared_ptr<Environment> &environment, const std::shared_ptr<CollisionDetection<dim>> &collision,
                      const std::shared_ptr<TrajectoryPlanner<dim>> &trajectory, const std::shared_ptr<Sampler<dim>> &sampler,
-                     const unsigned int attempts = 10, const double distance = 15);
+                     const size_t attempts = 10, const double distance = 15);
 
     Vector<dim> getSample() override;
 
@@ -59,7 +59,7 @@ template <unsigned int dim>
 GaussianSampling<dim>::GaussianSampling(const std::shared_ptr<Environment> &environment,
                                         const std::shared_ptr<CollisionDetection<dim>> &collision,
                                         const std::shared_ptr<TrajectoryPlanner<dim>> &trajectory,
-                                        const std::shared_ptr<Sampler<dim>> &sampler, const unsigned int attempts,
+                                        const std::shared_ptr<Sampler<dim>> &sampler, const size_t attempts,
                                         const double distance)
     : Sampling<dim>("GaussianSampling", environment, collision, trajectory, sampler, attempts), m_distance(distance) {
 }
@@ -74,7 +74,7 @@ template <unsigned int dim>
 Vector<dim> GaussianSampling<dim>::getSample() {
     Vector<dim> sample1, sample2, ray;
 
-    for (unsigned int count = 0; count < m_attempts; ++count) {
+    for (size_t count = 0; count < m_attempts; ++count) {
         sample1 = m_sampler->getSample();
         ray = m_sampler->getRandomRay();
         ray *= m_distance * m_sampler->getRandomNumber();
